requirements.txt
tensorflow-gpu==1.9.0
pandas==0.23.3
spacy==2.0.11
h5py==2.8.0